Workaround issue #100 in cc
cc parses function pointer declarations slightly incorrectly, which causes
function pointers that return typedefs to simply be typed as the typedef in
c-for-go. This change works around that bug by having the typedef name lookup
function not return the typedef name if the type is a function and wasn't
declared using a typedef.