Use the tag-mapped name when translating structs
Only one name is used to generate structs, but many names might be assigned to
a struct via typedefs. When a struct is used directly (as opposed to one of
its typedef forms) as a type in a parameter list, the generated function will
use the struct name, instead of the generated typedef name. This results in a
function that uses a type name that was never generated.

For example:
```c
struct foo {
    int bar;
};

typedef struct foo baz;

int qux(struct foo quux);
```
generates the `foo` struct under the `baz` name, but generates a `qux` function
whose parameter `quux`'s type uses the `foo` name.

This changelist solves this problem by using the same process when finding the
name of a parameter's type that generating the struct does.