Use the tag-mapped spec when creating a CGoSpec
Different types that share the same tag would still create different CGoSpecs.
This resulted in code that would reference variables that were never
generated.

For example:
```c
struct foo;

typedef int (*bar)(struct foo*);

typedef struct foo {
    int bar;
} foo;
```
would result in a pack helper that references `sizeOfStruct_fooValue` but the
generated variable would be `sizeOfFooValue` because of the typedef on th
struct.

Now the spec from the declaration found in the tag map is used to generate
parts of the CGoSpec so names will be consistent across all specs.