Fix parse error for empty constants
When an empty macro definition is encountered, such as:

```c
#define FOO
```

c-for-go would generate a constant whose name was equal to the printed
representation of `name`, which is a `[]byte`. So you would see
something like `[1 2 3 4 5]` (except ascii values) appearing in a
`go` source where an IDENT was expected.

Instead, drop it, since a const without a value isn't a very interesting const.