Treat long int types properly (long -> int32) to have exact bitsize.
Note that standards say "long guarantees at least 32 bits", which means we could use
just int and uint Go types for that. But on practice that's very platform dependent and most
code is using int32_t and uint32_t and users expect to have int32 and uint32 in Go bindings.

https://en.cppreference.com/w/cpp/language/types